An improved backwards migration for reverting from 7-area-unions to master
The previous version of the backwards part of migration 0006 could
lose geometries.  This version preserves all geometries that were
associated with a area, and unions them to minimize the number of
geometries required to represent the area.

Thanks to Jenny Duckett for helpful suggestions on how to rewrite
this so that it's shorter and clearer.